glibc 2.1 fixes.
The new glibc now has a utmpx interface, unfortunately this does not work out
of the box with hylafax(hylafax's utmpx calls also have problems on other
platforms).  As a temporary measure, hylafax will now default to regular utmp
on linux.

Also the hylafax configure is looking for some function prototypes in the wrong
include files.  This has been fixed.