glibc 2.1 fixes.
The new glibc now has a utmpx interface, unfortunately this does not work out
of the box with hylafax(hylafax's utmpx calls also have problems on other
platforms).  As a temporary measure, hylafax will now default to regular utmp
on linux.

Also the hylafax configure is looking for some function prototypes in the wrong
include files.  This has been fixed.


git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@60 5505949e-d877-4686-9e03-c53b7a51b376