Enhancement on Bug 583 	http://bugs.hylafax.org/bugzilla/show_bug.cgi?id=583
This implements proper desireddf handling in faxsend.  It also improves the
selection of DF faxq uses in preparation to hopefully minimize the work
faxsend may have to do when sending a image to remotes.

The desireddf of a job (and the destcontrols admin override) now place a
hard top limit on the best dataformat used for transmitting.


git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@828 5505949e-d877-4686-9e03-c53b7a51b376