waitNotifyPid is designed to keep the three "notifies" (notifyRecvBegun, notifyPageRecvd, and notifyDocumentRecvd) ordered.  As all of these notifies are not otherwise time sensitive, and as they can cause the protocol to hang, the code previously did a waitpid in a child process before running the next notify.  However, I don't think that this is acceptable (running waitpid in a child process), and so the waitpid (which maintains the order of the notifies) is now moved to the parent process.
This could potentially reintroduce protocol issues on heavily-stressed systems.  The proper solution
to this, I think, would be to run all three notifies in a common child process, but the difficulty
there is in passing the necessary function data from the parent process to the previously-begun child
process.  The usage of "sleep(1)" could be removed to further help speed up the child processing and
perhaps prevent protocol issues, but it has been the habit of HylaFAX code since long ago to use a
"sleep(1)" in order to "give the parent time".  This probably is really unnecessary and if really
needed could probably be done in other ways.  For now, though, I'm leaving it.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1189 5505949e-d877-4686-9e03-c53b7a51b376