This adds preliminary support for proprietary T.32 data format extensions on new MultiTech modem firmware.
MultiTech appears to have done this work in order to add support for JBIG.

This commit gets partially into the JBIG-on-Class-2 work, but a lot more
will need to get done in faxd/CopyQuality.c++ ... probably restructuring
some things to get the JBIG/JPEG parsing out from specific use in Class 1
ECM situations.

The reason this is being committed now, however, is to "fix" parsing of
the initial +FIS response so that MR and MMR can be sent to JBIG-supporting
receivers.  Otherwise HylaFAX doesn't know what to make of the "F" in the
DF section of the +FIS response and limits us to MH.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1327 5505949e-d877-4686-9e03-c53b7a51b376