This implements a "docrp" feature for Class1Modem::recvFrame for the intent of defining when using CRP is or isn't appropriate.
CRP may actually not be appropriate when receiving responses and may only be
appropriate when receiving commands.  Yet, it has proven useful to use in
receiving CFR/FTT and DIS.  However, it does us no benefit to use CRP when
waiting for a response to PPM/PPS, and so we will start by not using it
in those cases.

This may actually help reported situations where DCN is received in response
to CRP following PPS.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1343 5505949e-d877-4686-9e03-c53b7a51b376