Undo the quoted printable encoding of data in the dictionary file.
With the text encoded QP the character counts come out wrong and mess
up formatting.

Furthermore, as it has been pointed out on-list, the fact that it's encoded
QP does not prevent well-intending but bad-behaving text editors from
screwing up things still.  The best solution to the problem will be to
split the dictionary up into different language files...

... but I'm going to wait on doing that until the current gettext investigation
is closed.