* don't wait Class1TrainingRecovery when CRP was received (13 Jul 2006)
There's no need to wait at all, PLUS waiting that long could possibly
cause the sender to give up.

* fix error of sending a block twice, empty the 2nd time (13 Jul 2006)

Silly logic error causing blockFrame to be invoked after it just was.

* don't trigger hasV17Trouble for pages other than the first (13 Jul 2006)

If we're going to have true V.17 problems it will be in TCF or on the
first page and not in subsequent pages.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1367 5505949e-d877-4686-9e03-c53b7a51b376