tracePPM and tracePPR do essentially the same thing - the first was intended for tracing commands and the latter for tracing responses, however over time we've learned that we can get just about any signal at any time and so we've inevitably been putting all signals into both functions, and thus these two functions have been looking more and more the same with an ever increasing number of collisions to address
This change unifies tracePPM and tracePPR into traceFCF (which, in my opinion
is a long-overdue change).  Everything should work as before... the only
downside that I can tell, now, is that - as far as tracing goes - we're now
unable to distinguish between whether the expected signal was a command
or a response - but really, we weren't using the tracing feature this way,
anyway... and I don't see any value to tracing FCFs differently depending
on the expected siganl.