I'm coming to more and more of an opinion that we should be checking for modem responses during the entirety of Phase C data transmission - rather than just waiting until after all of the data is transmitted.  But, until we do that...
Some modems, like iaxmodem and Patton 2977s, are capable of detecting a
hangup, and so if that occurs during the middle of Phase C data transmission
the modem will respond/result with NO CARRIER before we ever get to end-of-data.
Consequently the remaining Phase C data that is sent to the modem can
potentially get interpreted as AT commands (if it coincidentally looks like
it) and will usually result in repeated ERROR messages.

We dont want those ERROR messages - or whatever the modem responses were
to the junk we sent to it - to be interpreted as responses to our subsequent
commands... so we need to flush the modem input.