This is somewhat experimental.
First off, getting ERROR in immediate response to AT+FRH=3 indicates
that we're on-hook.  No need to try AT+FRH=3 four more times before dying.

However, the check here for lastResponse != AT_NOCARRIER was essentially
for the same kind of thing:

  <-- AT+FRH=3
  --> NO CARRIER

But, the way that this is coded it would also catch on this kind of thing:

  <-- AT+FRH=3
  --> CONNECT
  --> NO CARRIER

... which technically may not mean that there was a disconnection, but could
merely mean that no HDLC data was received before the V.21 carrier dropped.

So, the idea is that in either of the above NO CARRIER situations that on
the next loop we'll see this:

  <-- AT+FRH=3
  --> ERROR

and then we know for-sure that a disconnection occurred.  If we don't see
that happen then we may need to further enhance Class1Modem::recvFrame to
better detect "eot" situations (with NO CARRIER).



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1373 5505949e-d877-4686-9e03-c53b7a51b376