Let's say that in receive mode this happens:
  --> Phase C ECM data
  --> NO CARRIER
  --> PPS-NULL
  <-- MCF
  --> strange V.21 signal (MCF, CFR, NSF, etc)

Then we'll have started a decoding child process, but the strange signal causes
a fallback to non-ECM behavior and (the error) is that we forget to take care
of the child.

This patch fixes that error of forgetting the child.

It also fixes an error where the same kind of thing would occur when
SaveUnconfirmedPages is set to false.  (In fact, that config option should probably
be removed entirely and always have the behavior in practice ... otherwise getting
the code right for both is a rather tricky scenario for so little benefit.)

The patch also fixes the saving of the page when the strang signal condition takes
place.

The patch also (finally) simplifies the repetitive coding for this abort procedure
into a function.