this replaces Class1TrainingRecovery with Class1SwitchingCmd
waiting a fixed period of time is an imperfect solution to simply trying
to avoid transmitting at the same time as the other end - using +FRS seems
to be a much more direct approach to that end

as for the modems that don't support +FRS ... well, I'm truly sorry.  In
any case 70 ms wait is not likely to be any worse than 1500 ms.