this replaces Class1TrainingRecovery with Class1SwitchingCmd
waiting a fixed period of time is an imperfect solution to simply trying
to avoid transmitting at the same time as the other end - using +FRS seems
to be a much more direct approach to that end

as for the modems that don't support +FRS ... well, I'm truly sorry.  In
any case 70 ms wait is not likely to be any worse than 1500 ms.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1384 5505949e-d877-4686-9e03-c53b7a51b376