After publicly announcing my intent in doing this there was at least one supportive (private) response and no objections at all to moving to 5.0.0. This is done primarily to highlight the difference that exists between these releases and those at hylafax.org.
Originally the intent was that these would be sub-releases (pre-released
select CVS snapshots named "sub-release"), but that was apparently
objectionable to some at hylafax.org.  So the next step was to maintain
a different code branch entirely (so as to remove the objectionable phrases
such as a perceived sanction by "The HylaFAX Development Team" from the
repository) and to port changes here to there as time goes on.

That porting work was tedious and unfulfilling, and when those at hylafax.org
picked up on doing it I was quite pleased and felt liberated from doing
that work myself.  However, they haven't been consistent or thorough in what
work they've been taking - missing some bugfixes - missing some protocol
fixes - missing some feature additions - but at the same time taking others.
There's been very, very little (if any) discussion from hylafax.org developers
with me about developments here... and so I can only suppose that it will
continue this way.  Inevitably they're going to pull and port what changes
suit them, but they're obviously going to leave some changes alone without
discussing them at all.  And so that basically means that this code will
forever remain truly forked from that code.

That's okay, really.  It's fine.

So it's now time to remove the old "sub-release" versioning system, because
it's a relic of an ambition that will not happen... and it's confusing to
people (at best) or (worse) even misleading.