The ECM HDLC frame decoder would read the bitstream and would pull out ECM frames from there.
Most of this change is directed towards removing the requirement that the
HDLC frame be immediately followed by an HDLC sync flag.  That way if the
data corruption occurs on the flag and not in the frame itself then the
data is still useable.

The way in which I've coded this, however, in merely counting the frame
size as it builds allows the decoder to recover from data corruption more
quickly than before.  So that's an added bonus, I guess. :-)



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1484 5505949e-d877-4686-9e03-c53b7a51b376