This is my first step towards eventually fixing job groups.
This enables sendfax to submit a job group using -z or multiple -d options.

Indeed, I suspect that the code-as-was represented a bug.  The protocol
is very deliberately designed for a succeeding job to inherit values from a
preceding one in the same client session.  The "JOB DEFAULT" prevents
that inheritance from occurring and inhibits job grouping from occurring.

In turn, sendfax really shouldn't be repeating job parameters that are
identical to those in the preceding job in the group... it will save
on communication and will make the sendfax protocol usage more elegant.  This
is something to do later.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1494 5505949e-d877-4686-9e03-c53b7a51b376