DesiredDF was being trumped by RTFCC in faxsend.  I think that it's clear in a case where desireddf is provided on the command-line to faxsend that faxsend needs to reign-in RTFCC behavior in that case to accomodate the desireddf setting.
Furthermore faxq ignored DesiredDF in its document preparation (as it does
with the other Desired* items).  It really does need to prepare the document
according to the DesiredDF value.

The consequence of both of these together amounts to it being impossible
for JobControl to control the DF used.

This change addresses both faxq and faxsend.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1508 5505949e-d877-4686-9e03-c53b7a51b376