PAM support was a bit messed up.
Required reading: The Linux-PAM Application Developers' Guide (Aug 2006)

HylaFAX needs to be able to authenticate against either PAM, hosts.hfaxd, or both.
And because of the way that traditional hfaxd logins can work (even without
password authentication) we have to make two runs by pam_authenticate, and we
cannot leave all of the authentication up to PAM.  We just can't impose a PAM
dependency on the software in general.

The first run by pam_authenticate is for authentication without passwords.  This
can be granted by PAM as well as hosts.hfaxd.

The second run by pam_authenticate is for authentication with the password.

There were two main problems with the way that PAM authentication was being done
before.

First, we were trying to do both pam_authenticate runs with the same
PAM session, and in-between those two runs we were altering PAM_CONV.  This may
have been possible in days gone by or with Solaris or BSD, but with Linux this
is not acceptable, at least when the first pam_authenticate results in
PAM_INCOMPLETE... in that case it's like tearing the rug out from under the PAM
module.  So the solution to this was to do the two different pam_authenticate runs
with two different PAM sessions.

Second, PAM modules such as pam_unix require privileged access to the password
databases such as /etc/passwd and /etc/shadow.  The fix for this is to bump our
euid temporarily for the PAM interaction.

In addition to these fixes I added a lot of error reporting enhancements and
changed a couple of things (like passwd -> passWd to avoid potential naming
collissions) that I felt were appropriate.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1509 5505949e-d877-4686-9e03-c53b7a51b376