These changes make faxmail not suck so bad.
The bulk of the matter revolves around how externally-formatted
parts are handled and changing it to better-suit common tools
and user intent.

Originally it appears that faxmail intended to use externally-
formatted mail parts in exactly the same way as it handles parts
that are formatted internally, and that is in-line.  So it
expected the external formatters to feed back Postscript snippits
that were somehow, magically formatted properly for the in-line
Postscript state.  Actually getting an external formatter that
could conceivably do this well in the variety of states that
the in-line Postscript could be in would be a difficult at best.
Existing coversion tools (pdf2ps, tiff2ps, etc) almost always
produce full-sized Postscript pages that are independent and
are not appropriate for in-line insertion into faxmail output.

So, as I saw it, there were basically two options:

1) build-in a Postscript interpreter into faxmail... at least
   enough of one to parse and filter the external formatter
   output from common tools so that it could place the output
   in-line as appears to be originally intended.

2) add a page break before adding external part formatting and
   allow the externally-formatted part to be displayed as the
   common tools expect

I chose to go with the latter option.  Not only is it an easier
solution to implement, but from many years of seeing people
complain about faxmail behavior and seeing their HOW-TO documentation
on MIMEConerters usage (search the archives for "faxmail" and
"showpage") it appears that this latter option is exactly what
users expect as well.  In fact, they tend to interpret the
"showpage misplacement" as a bug.

So I merely add a page break just before adding externally formatted
parts.  I also keep track of things so that a second trailer is not
added if the last part is externally formatted.

This should remedy some complaints that faxmail produces blank
pages and misses some MIME parts.

I've added application/pdf, image/tiff, and application/octet-stream
(in order to handle Microsoft mails that only use the filename
extension in identification of attachment MIME type) MIMEConverters
by default.  This should handle the frequent complaints that
faxmail does not support PDF, etc.

I've also modified faxmail to support MS-DOS formatted (CR+LF
line endings) mails.  This may be somewhat contrary to spec,
but I do believe that it will ultimately prove to be, again,
what users expect.  This comes at the expense of the risk that
in a quoted-printable part that - if an intentional CR comes
right before the line ending (LF) - the CR will be ignored.
If this becomes a problem it can be addressed by causing faxmail
to recognize whether CR+LF is being used from the envelope
headers.

I've added a faxmail -N "formatEnvHeaders" feature to allow the
administraor to force suppression of the formatting of the
headers.  (Previously this was only accomplished by adding an
X-Fax header to "clear" the headers.)