Sometimes echo can be so bad on a channel that we will hear ourselves back, especially after signalling something like MCF or CFR that is short (in contrast to PPR which is long).
Although this situation really represents a medium problem - something's wrong
on the call itself - I think that we need to do what we can to address it
where we can... tolerance as best as possible.

So this is a first stab at that.  This only does it in Phase C ECM, so there
are other places where this needs to be addressed (although they would be much
more infrequent elsewhere) if this helps.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1573 5505949e-d877-4686-9e03-c53b7a51b376