Adjusting killtime without the explicit instruction of the user is plain wrong.
It was easy to reproduce with delayJob killtime adjustment, but it could have
happened in other ways, too, that a delayed job tts would end up falling very
near or on the killtime, and (especially when they were at the same time) then
we'd invoke runJob and timeoutJob at the same time... which really is a
contradiction.

So firstly, let's stop messing with killtime in delayJob.

Secondly, tell timeoutJob (in setDead) to turn off the tts timer and likewise
tell runJob to not bother if the job is dead already.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1584 5505949e-d877-4686-9e03-c53b7a51b376