This is essentially straight from hylafax.org bugzilla 723.
For some reason I didn't notice the display offset, that I'm sure was there.
The incongruency between clients and hfaxd was fixed, but not the display.
So this fixes the display.

The reason that this wasn't committed a year ago is because this wording
was unintelligible to me:

     * We want the actual offset of the local now from UTC,
     * This means we have to "force" daylight savings off, otherwise
     * the offset we get back is of the real timezone (the time
     * functions automatically compensate if DST is set).

Although this gives the practical reasoning, it doesn't explain *why*
tm_isdst needed to be set to 0, and so I worried about making a change
that didn't make sense at the time.  Here's my take on it:

     * mktime() gives us the seconds elapsed since the
     * beginning of calendar time.  And so in a
     * "daylight savings" condition the clock shows
     * a time ahead of calendar time - where that
     * amount of time has not elapsed yet - and thus
     * subtracting mktime(&tm) from mktime(&gmt) will
     * not immediately give us the clock offset
     * between local time and GMT when DST applies.  To
     * compensate we set tm.tm_isdst to 0 always, and
     * this will force mktime() to believe that the
     * those seconds have elapsed, and thus we'll get
     * an accurate gmtoff value.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1607 5505949e-d877-4686-9e03-c53b7a51b376