Extensive patch to change all error status notices to include an errorcode.
Error codes are described in faxd/README.errorcodes.  They are delimited in
the status notices by {brackets} and are of the form "E000".

The errorcode is stored in the q-file and clients are allowed to query for
it.

This should allow clients to customize the messages keyed from the error
codes.

Furthermore, it should facilitate the futre translation of these messages in
mailings, as well.