From Aidan:
Fix a bug where the modem is "released" in the middle of the batch.  If the
preparation failes, and there are other jobs in the batch (before or after), we
need to take care that delayJob/deleteJob don't release the modem, because we
need it for the rest of the jobs.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1633 5505949e-d877-4686-9e03-c53b7a51b376