This is the send-portion of the earlier +FDB=1 debugging support work.
This implements ModemDoPhaseCDebug config option which will make HylaFAX
query the modem during Phase C transmit to check for modem responses.  This
is done at that time so that they can be displayed at roughly the correct
time in the log, rather than being buffered up for the end.

This is left off by default, although in theory it should be harmless enabled
all of the time.  There is some potential risk if something goes wrong... like,
say the modem shows NO CARRIER before the DTE signals <DLE><ETX> (i.e. I can
envision a modem like iaxmodem doing this if there is a premature hangup).  And,
in that case the NO CARRIER response will get taken from the buffer and simply
logged rather than interpreted... which means that we have some work to do
in order to make this better... but this should be good for now.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1667 5505949e-d877-4686-9e03-c53b7a51b376