Some modems appear to return ERROR to AT+FRH=3 when the modem is *not* on-hook. So to persist through this condition we are setting up the Class1HookSensitivity feature which will permit the modem to continue through those instances.
I'm only applying it now to the DCS-receive portion now.  I'm sure later
it will get applied to other places where "gotEOT" is handled.

I'm seeing the condition happen with a Zyxel Omni 288S modem when it hears
persistent CNG after answering and sending prologue:

<-- AT+FRH=3
(5 seconds or so elapse)
--> ERROR

This is most likely because the V.21 DSP picks up on the CNG audio, but does
not issue a CONNECT due to not seeing HDLC.  When the CNG audio drops then
the modem gives ERROR.  It probably should not be doing anything of the sort,
merely waiting until HDLC is detected in a carrier, but if it were to follow
some interpretations of the T.31 sample sessions, then maybe a NO CARRIER
result would have been more appropriate than ERROR.

In any case, setting Class1HookSensitivity to some number like 10 should
prevent the situation from causing the Class 1 driver to give up.