Basically there were two problems here... first, Sys::now() gives us seconds, and conf.answerResponseTimeout is in milliseconds.  We can't compare the two directly as they're not the same units.  Otherwise we'll end up waiting through 50 hours of RINGs or whatever other steady stream of responses the modem may try to send our way.
Second, if we do see multiple RINGs after the answer command (i.e. ATA) then
it means that the modem somehow didn't get the message, and we need to try
sending the command over again.

Note that there is a risk that between rings the call switched (e.g. the first
hung up and the second began) and so we could get the wrong CallID information
on the latter call.  Unfortunately, this is an inherant risk in analog
and analog-simulating environments.  Detecting call begins and ends is nearly
impossible to be completely sure all of the time.