When a continuation cover page was used the pagination would get screwed up (totpages would grow by one for the added page)... and this causes confusion for the receiver (better-off without the continuation cover).
This patch addresses that.  It's done by implementing a "nocountcover"
feature that gets used automatically for continuation cover pages.
However, the feature is also availble to the client so that the client
can (as has been requested in days gone by) prevent the traditional
cover page as submitted by the client from being paginated with the rest.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1762 5505949e-d877-4686-9e03-c53b7a51b376