This removes all of the "lastbyte" code for sending MMR use as well as removes MMR support from correctPhaseCData.
Long ago I had made an incorrect conclusion that all 2D code words ended with a
"1".  And to follow the pattern in the MH and MR contexts sending RTC... in
order to send a properly-encoded EOFB I therefore used the "lastbyte" code and
such to determine exactly how many zero bits needed to lead the EOFB signal.
And because of the incorrect initial conclusion about 2D code words it thus
meant that whenever an MMR image ended with a 2D code word ending with a "0"
that the EOFB would be corrupt through sendRTC().

Our MMR encoder puts EOFB correctly at the end of the data stream.  There is no
need to try to second-guess our own MMR encoder.  Since we're sticking a tagline
at the top of every page (virtually, only not when the page is a very, very thin
strip) it means that EOFB will be properly there because when we do that we must
reprocess the entire MMR image.

This is different with MH and MR because in those cases we're not re-encoding
the entire page.  And thus a faulty RTC signal (or a duplicate RTC) in the image
data can creep in from the source TIFF file which supposedly can be corrupt from
Ghostscript or from the client.  So with MH and MMR we are wise to use
correctPhaseCData to try to ensure a proper RTC; in the case of MMR it's simply
a waste: there's nothing to correct that hasn't already been corrected by the
same encoder either in tagline generation or in soft-RTFCC.

So this ultimately will fix a few cases, perhaps rare, where the receiver was
signalling DCN after PPS.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1805 5505949e-d877-4686-9e03-c53b7a51b376