This causes the fmtTime() time format to always include hours.
Logs like this:

JOB 1086 (sleeping dest +1234 pri 126 tts 4:58 killtime 4:58:00): SLEEP FOR 4:58

Just aren't that intuitive with respect to the "4:58".  To some readers that
may normally mean 4 minutes and 58 seconds... but to others it may look like
four hours and 58 minutes.

One such "reader" is MySQL where "4:58" will be interpreted as nearly five hours
instead of nearly five minutes.  In order to communicate the proper time one must
use "0:04:58".

So to interface things one constantly needs to parse the fmtTime()
output to add the hours (if not present).  Let's save everyone a headache and
simply always include hours to the fmtTime() time format.

JOB 1086 (sleeping dest +1234 pri 126 tts 0:04:58 killtime 4:58:00): SLEEP FOR 0:04:58

That's a lot more clear to understand.