These are some enhancements/fixups for Mac OS X Leopard (darwin9).
Of import is to note that a "-1" define, as for _POSIX_PRIORITY_SCHEDULING
indicates that it is unavailable... and thus should mean the same thing as
leaving it undefined.  Our configure "checkForDefined" does not anticipate
this.  (And it probably should.)  But for the sake of an easy fix here
without potential repercussions for other systems that may use a -1 define
otherwise... we'll just continue to address the issue with darwin9 as we
did with darwin8.  See:

http://lists.apple.com/archives/Unix-porting/2005/Jul/msg00027.html



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1828 5505949e-d877-4686-9e03-c53b7a51b376