Users who wish to use other than ASCII characters in their taglines have been met with the impossible task of customizing their taglines to use their characters.  This has been because of the inability for the tagline imager to reference glyphs numbered beyond 255.  This was limited twofold:
1) because of restrictions/assumptions inside of PCFFont.c++ itself, and

2) because there was no way for PCFFont.c++ to understand a multibyte
   format string.

So this patch adds an attribute to fxStr types, "isutf8" which indicates
whether or not the string is UTF-8 encoded.  This is done currently for the
sole reason of making the "next" function work as expected... although it
probably would have worked anyway because in UTF-8 all "encoded" bytes are
high-bit, and I doubt that we use any next function calls on high-bit bytes.
However, I added the attribute anyway.

So the patch adds a modem configuration option "TagLineLocale".  This option
is used twofold:

1) as an indicator for whether or not the TagLineFormat is encoded in UTF-8, and

2) to call setlocale so that the strftime function will operate in a locale-
   specific way.

Of course, HylaFAX's default font only has ASCII characters in it, and so
anyone using UTF-8 encodings in TagLineFormat will need to get a unicode PCF
font in order for this to be useful.