We've been telling people for years now that the client is where the sorting should take place.  Yet, we don't see most client developers doing it.  Even more telling is that we're not even doing it ourselves in faxstat... instead telling people to pipe faxstat output through external sort mechanisms.
Obviously sorting takes CPU.  And obviously if there are 10,000 jobs
in the queues and there are hundreds of people sorting the queues
at the same time, then there's going to be a heavy price paid for it.

Well, first... I wonder how plausible the costly scenario really is.
Who has 10,000 jobs in the queue and has hundreds of people querying
simultaneously without adequate CPU to handle the sorting and with
an expectation that responsiveness should be quick?  I think it's
an implausible scenario.

So, I think that it's time to bite the bullet and sort.

This change implements basic sorting.  It doesn't sort the doneq/sendq
numerically (which just would require a numerical comparison function),
but it should be significantly better than no sorting at all.