Adds PollLockPokeModem
When polling for the presence/removal of a UUCP lockfile, also test to make
sure the modem is still reachable.  This is useful for scenarios where the
modem is apt to become disconnected while the modem is idle, such as an
external modem that may lose power, where the administrator wants notification
of this (as a ‘‘wedged’’ condition).