This makes faxmail decode headers that are encoded with RFC-2047.
It does *not*, however, add any capability to faxmail in switching character
encodings as indicated in the RFC-2047 part.  Thus the resulting postcript
will *still* use the "standard" encoding which is basic, 7-bit ASCII.

Thus, although this will decode RFC-2047 headers it will not, alone, make
high-bit characters (such as umlauts and such) available.  More work will
need to be done - Poscript work - in order to accomplish that.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1908 5505949e-d877-4686-9e03-c53b7a51b376