This does a couple of things...
1) It makes faxq check some real basic things about FIFO messages before
   processing them blindly such that an fxStr assertion could fail.

2) It prevents faxq from self-corrupting the FIFO messages itself.  faxq
   was corrupting FIFO messages when the read() exceeded the buffer size.
   So if the FIFO activity was heavy, then faxq could possibly corrupt
   FIFO messages.  (I don't think that a FIFO overflow is possible, as it
   seems that it will simply block instead.)

However, it does not bullet-proof faxq FIFO message handling.  You can get
faxq to get into trouble if you send it lots of garbage.  Like, dump any
binary file data into the FIFO, and you can get faxq to get locked up or
into endless loops.  Debugging all of that would be quite an effort, and
it may be unwarranted, since the FIFO-writing applications such as faxgetty
and faxsend should not be dumping that kind of data in there anyway.

This work was not intended to address bullet-proofing, but it was intended
to address the cases where the FIFO messages were corrupted by improper
parsing.