This is a somewhat experimental change which adds a couple of new features:
"ignoremodembusy" q-file entry which is the client-set request to send the
job regardless of any existing lock file or busy modem state

"AllowIgnoreModemBusy" faxq config file entry to be able to permit the above

The intent of this is to allow an incoming call to be able to be answered and
yet to send a fax to the caller.  This will happen, for example, if a
person calls in from a phone on a fax machine, wades through some kind of IVR
system to retrieve a fax, and then presses "Start" or "Answer" on their
fax machine once the CNG beeps start.

Technically this procedure is contrary to T.30 as this is not fax polling.
It's basically the opposite of the "dialAnswer" feature that has existed in
the faxanswer tool for a long time now.

A sample implementation using IgnoreModemBusy would be to have a DynamicConfig
script that looks like this:

#!/bin/sh
CALLERID=$3
MODEM=`echo $1 | sed 's/^\/dev\///g'`
echo test | sendfax -n -w -t 1 -T 1 -i answersend -O ignoremodembusy:yes -h $MODEM\@ -d $CALLERID
echo "RejectCall: true"
exit 0

... and then there is a JobControl script that looks like this:

#!/bin/sh
. etc/setup.cache
. bin/common-functions
QFILE=sendq/q$1
parseQfile
if [ "$jobtag" = "answersend" ] && [ "$ignoremodembusy" = "1" ]; then
    echo "ModemDialCmd: \"ATH1\\nATD\""
    echo "ModemSoftResetCmd: AT"
fi
esac
exit 0



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1990 5505949e-d877-4686-9e03-c53b7a51b376