This has been a long-standing issue that, for some reason, wasn't fully addressed by hylafax.org bug 445 and was possibly only made less-likely due to hylafax.org bug 443.
If faxqclean either isn't being run, isn't working properly, or isn't
cleaning up docq image files quickly enough... AND... if faxq does not
clean up prepared image files after the job completes or fails (for some
reason, i.e. malfunction or system crash or ... something)... then it's
entirely possible (and easy to reproduce) that faxq will re-use the old
image file from an old job when the docq seq number comes 'round again.

This change now makes a deliberate attempt to remedy this by comparing
the modification time on the source image file with that of the target
image file and ensures that the target is newer than the source before
reusing the target.  If it is older then it will delete the target before
recreating it.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@1998 5505949e-d877-4686-9e03-c53b7a51b376