From Aidan @ hylafax.org...
hfaxd/faxq: Add -c <option> to hfaxd/faxq as done in faxsend/faxgetty

This means all the "server" components of HylaFAX take -c option
arguments on the command line.  faxq, faxsend and faxgetty have
already done this, this gives hfaxd that ability too.  It's useful
if you need to start hfaxd with some option that you don't want
going in the config file.

This also documents the -c option in the rest of the man pages.