From Aidan @ hylafax.org...
PageHandling:  Fix handling of large blank areas

If the blank data at the end of a page is more than 64K, then it doesn't fit
into a 'Z%04x' field.  In reality, the only way you get more than 64K of blank
data is if the tiff encodes the blanks pixel for pixel, and *not* encoding.
Fortunately, the only program I've seen that does that is Microsoft FAX
service.

Unfortunately, if it is > 64K, we need to limit it, otherwise we end up
something like:
	Internal botch; unknown post-page handling string "e1Z131b0P"