This adds the "usecolor" feature to the fax job req files and adds an -O usecolor:[yes|no|only] feature to sendfax and to the hfaxd protocol.
(usecolor:only is to indicate that the fax is to be prepared only in color,
and if the receiver does not support color then the job fails.)

It goes as far as getting the information passed to the *2fax scripts.

Next will be to get those scripts formatting in color (and simultaneously
in monochrome, when appropriate)... and then to finally get faxsend to
actually handle the color scenario.