From Aidan @ hylafax.org...
DialRules: Avoid abort on zero length matches

Pattern rules may contain 0 length matches, such as:
   (^[+])?([0-9]*)

We have nothing to replace it with (mlen = 0), so trying to extract
it from the string causes and abort.  If there is nothing to replace
it with, there is nothing to try extracting for insertion.

If the match reference was not matched in the string, StartOfMatch and
EndOfMatch both return -1, leaving mlen = 0, making it an ideal check
if there is something to extract and insert.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@2175 5505949e-d877-4686-9e03-c53b7a51b376