From Aidan @ hylafax.org ...
hfaxd: Release old accept fd

By unlinking from the dispatcher early, we prevented it from being
closed in the child, leaving an open listening fd on the main hfaxd
port.

If the main (parent) hfaxd exits, the child will inheret the listen
but won't actually include it in it's selects, or handle it.  It just
prevents a new master hfaxd from binding to it.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@2182 5505949e-d877-4686-9e03-c53b7a51b376