From Dennis Jenkins...
Added member variables "pasv_min_port" and "pasv_max_port" to BOTH
InetFaxServer and InetSuperServer (was required in InetSuperServer due to
what exists at the time that the command line is parsed.  I opted to follow
the existing convention instead of just shoving them into two globals.)

Added '-p' option to getopts, and logic to validate that the given port
range is sane.