The changes made in revision R2258 seem to have been overzealous.
FaxQueuer expects the sendq file to remain in-place for longer or we'll get "job file vanished" complaints.  Furthermore, a report seems to allude to this change being the catalyst to faxq crashing when the conditions are right.

This modifies the changes from R2258 to function as before except with the additional logging provided by R2258.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@2329 5505949e-d877-4686-9e03-c53b7a51b376