This removes CONFIG_FIFOBUG and CONFIG_OPENFIFO workarounds for a sanctioned approach.
This was a result of a conversation on hylafax-users@hylafax.org between Daren Russell, Konstantin Belousov, and myself.  Daren was encountering problems with fast-looping in faxgetty on select() and read() calls.  This appears to happen do to a change in FreeBSD on how select() works.

Looking into things, it appears that HylaFAX has been relying on a bug in the select() implementation as a feature to avoid EOF detection.  Over the years as various UNIXes have dealt with this matter HylaFAX has apparently attempted to cope by relying on other undefined behaviors (such as read+write on a FIFO).

The sanctioned approach to avoid EOF is to have the reading process additionally open up the FIFO in write-only... and then do nothing with it other than leaving it open.

https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=166071
https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=94772
http://fixunix.com/unix/350803-use-select-fifo-after-has-closed.html